Refactor apcd_membership step userlink function into project_membership step
